User-Action/requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
fastapi==0.115.6
frozenlist==1.5.0
h11==0.14.0
idna==3.10
kafka-python==2.0.2
kafka-python-ng==2.2.3
multidict==6.1.0
nats-py==2.10.0
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2025.2
pymongo
redis==5.2.1
requests==2.32.3
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
starkbank-ecdsa==2.2.0
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.33.0
websocket-client==1.8.0
websockets==13.1
yarl==1.15.2
